pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "ocalcli"
dynamic = ["version"]
requires-python = ">= 3.9"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Eric Davis" },
  { name = "Brian Hartvigsen" },
  { name = "Joshua Crowgey" },
]
maintainers = [
  { name = "David Barnett" },
  { name = "Martin Dengler" },
]
description = "Outlook Calendar Command Line Interface"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "msal",
  "httpx",
  "typer",
  "rich",
  "python-dateutil",
  "icalendar",
  "platformdirs",
  "pydantic",
  "toml",
]

[project.urls]
Repository = "https://github.com/insanum/ocalcli"
Issues = "https://github.com/insanum/ocalcli/issues"
Changelog = "https://github.com/insanum/ocalcli/blob/HEAD/ChangeLog"

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "black",
  "mypy",
  "types-python-dateutil",
  "types-requests",
]

[tool.setuptools]
packages = ["ocalcli"]

[tool.setuptools.data-files]
"share/man/man1" = ["docs/man1/ocalcli.1"]

[tool.setuptools_scm]
version_file = "ocalcli/_version.py"

[project.scripts]
ocalcli = "ocalcli.cli:app"

[tool.ruff]
line-length = 80
extend-exclude = ["tests/cli/"]

[tool.ruff.lint]
# Enable Errors, Warnings, Flakes
select = ["E", "W", "F"]

[tool.ruff.format]
# Permit mixed quote style, project currently uses a mix of both.
quote-style = "preserve"

[tool.ruff.lint.extend-per-file-ignores]
"*.pyi" = ["E501"]

[tool.mypy]
mypy_path = "ocalcli:stubs:tests"
